--- Halcyon.Api/Features/Account/ForgotPassword/ForgotPasswordDtos.cs
@@ -1,3 +1,0 @@
-namespace Halcyon.Api.Features.Account.ForgotPassword;
-
-public record ForgotPasswordRequest(string EmailAddress);
