+++ Halcyon.Api/Features/Account/Login/LoginResponse.cs
@@ -1,5 +1,3 @@
 ﻿namespace Halcyon.Api.Features.Account.Login;
 
-public record LoginRequest(string EmailAddress, string Password);
-
 public record LoginResponse(string AccessToken);
