+++ Halcyon.Api/Features/Account/Register/RegisterRequest.cs
@@ -1,11 +1,19 @@
 ﻿using FluentValidation;
 using Halcyon.Api.Common.Validation;
 
-namespace Halcyon.Api.Features.Users.CreateUser;
+namespace Halcyon.Api.Features.Account.Register;
 
-public class CreateUserRequestValidator : AbstractValidator<CreateUserRequest>
+public record RegisterRequest(
+    string EmailAddress,
+    string Password,
+    string FirstName,
+    string LastName,
+    DateOnly DateOfBirth
+);
+
+public class RegisterRequestValidator : AbstractValidator<RegisterRequest>
 {
-    public CreateUserRequestValidator(TimeProvider timeProvider)
+    public RegisterRequestValidator(TimeProvider timeProvider)
     {
         RuleFor(x => x.EmailAddress)
             .NotEmpty()
@@ -13,14 +21,9 @@ public class CreateUserRequestValidator : AbstractValidator<CreateUserRequest>
             .MaximumLength(255)
             .WithName("Email Address");
 
-        RuleFor(x => x.Password).NotEmpty().MinimumLength(8).MaximumLength(50).WithName("Password");
+        RuleFor(x => x.Password).NotEmpty().MinimumLength(8).MaximumLength(50);
         RuleFor(x => x.FirstName).NotEmpty().MaximumLength(50).WithName("First Name");
         RuleFor(x => x.LastName).NotEmpty().MaximumLength(50).WithName("Last Name");
         RuleFor(x => x.DateOfBirth).NotEmpty().InThePast(timeProvider).WithName("Date Of Birth");
-
-        RuleForEach(x => x.Roles)
-            .Must(role => Roles.AssignableRoles.Contains(role))
-            .WithMessage("Role '{PropertyValue}' is not recognized.")
-            .WithName("Roles");
     }
 }
