+++ Halcyon.Api/Features/Account/Register/RegisterResponse.cs
@@ -1,0 +1,3 @@
+﻿namespace Halcyon.Api.Features.Account.Register;
+
+public record RegisterResponse(Guid Id);
