+++ Halcyon.Api/Features/Account/ResetPassword/ResetPasswordRequest.cs
@@ -2,6 +2,8 @@
 
 namespace Halcyon.Api.Features.Account.ResetPassword;
 
+public record ResetPasswordRequest(Guid Token, string EmailAddress, string NewPassword);
+
 public class ResetPasswordRequestValidator : AbstractValidator<ResetPasswordRequest>
 {
     public ResetPasswordRequestValidator()
