+++ Halcyon.Api/Features/Account/ResetPassword/ResetPasswordResponse.cs
@@ -1,5 +1,3 @@
 ﻿namespace Halcyon.Api.Features.Account.ResetPassword;
 
-public record ResetPasswordRequest(Guid Token, string EmailAddress, string NewPassword);
-
 public record ResetPasswordResponse(Guid Id);
