+++ Halcyon.Api/Features/Profile/ChangePassword/ChangePasswordResponse.cs
@@ -1,5 +1,3 @@
 namespace Halcyon.Api.Features.Profile.ChangePassword;
 
-public record ChangePasswordRequest(string CurrentPassword, string NewPassword, uint? Version);
-
 public record ChangePasswordResponse(Guid Id);
