+++ Halcyon.Api/Features/Profile/DeleteProfile/DeleteProfileRequest.cs
@@ -1,5 +1,3 @@
 ﻿namespace Halcyon.Api.Features.Profile.DeleteProfile;
 
 public record DeleteProfileRequest(uint? Version);
-
-public record DeleteProfileResponse(Guid Id);
