+++ Halcyon.Api/Features/Profile/DeleteProfile/DeleteProfileResponse.cs
@@ -1,0 +1,3 @@
+﻿namespace Halcyon.Api.Features.Profile.DeleteProfile;
+
+public record DeleteProfileResponse(Guid Id);
