+++ Halcyon.Api/Features/Profile/UpdateProfile/UpdateProfileResponse.cs
@@ -1,0 +1,3 @@
+namespace Halcyon.Api.Features.Profile.UpdateProfile;
+
+public record UpdateProfileResponse(Guid Id);
