--- Halcyon.Api/Features/Users/CreateUser/CreateUserDtos.cs
@@ -1,12 +1,0 @@
-﻿namespace Halcyon.Api.Features.Users.CreateUser;
-
-public record CreateUserRequest(
-    string EmailAddress,
-    string Password,
-    string FirstName,
-    string LastName,
-    DateOnly DateOfBirth,
-    List<string>? Roles
-);
-
-public record CreateUserResponse(Guid Id);
