+++ Halcyon.Api/Features/Users/CreateUser/CreateUserResponse.cs
@@ -1,0 +1,3 @@
+﻿namespace Halcyon.Api.Features.Users.CreateUser;
+
+public record CreateUserResponse(Guid Id);
