+++ Halcyon.Api/Features/Users/DeleteUser/DeleteUserRequest.cs
@@ -1,0 +1,3 @@
+﻿namespace Halcyon.Api.Features.Users.DeleteUser;
+
+public record DeleteUserRequest(uint? Version);
