+++ Halcyon.Api/Features/Users/DeleteUser/DeleteUserResponse.cs
@@ -1,5 +1,3 @@
 ﻿namespace Halcyon.Api.Features.Users.DeleteUser;
 
-public record DeleteUserRequest(uint? Version);
-
 public record DeleteUserResponse(Guid Id);
