+++ Halcyon.Api/Features/Users/LockUser/LockUserRequest.cs
@@ -1,5 +1,3 @@
 ﻿namespace Halcyon.Api.Features.Users.LockUser;
 
 public record LockUserRequest(uint? Version);
-
-public record LockUserResponse(Guid Id);
