+++ Halcyon.Api/Features/Users/LockUser/LockUserResponse.cs
@@ -1,0 +1,3 @@
+﻿namespace Halcyon.Api.Features.Users.LockUser;
+
+public record LockUserResponse(Guid Id);
