+++ Halcyon.Api/Features/Users/UnlockUser/UnlockUserRequest.cs
@@ -1,5 +1,3 @@
 ﻿namespace Halcyon.Api.Features.Users.UnlockUser;
 
 public record UnlockUserRequest(uint? Version);
-
-public record UnlockUserResponse(Guid Id);
