+++ Halcyon.Api/Features/Users/UnlockUser/UnlockUserResponse.cs
@@ -1,0 +1,3 @@
+﻿namespace Halcyon.Api.Features.Users.UnlockUser;
+
+public record UnlockUserResponse(Guid Id);
