--- Halcyon.Api/Features/Users/UpdateUser/UpdateUserDtos.cs
@@ -1,12 +1,0 @@
-namespace Halcyon.Api.Features.Users.UpdateUser;
-
-public record UpdateUserRequest(
-    string EmailAddress,
-    string FirstName,
-    string LastName,
-    DateOnly DateOfBirth,
-    List<string>? Roles,
-    uint? Version
-);
-
-public record UpdateUserResponse(Guid Id);
