+++ Halcyon.Api/Features/Users/UpdateUser/UpdateUserResponse.cs
@@ -1,0 +1,3 @@
+namespace Halcyon.Api.Features.Users.UpdateUser;
+
+public record UpdateUserResponse(Guid Id);
